requirements.txt
pytest~=8.0.0
requests~=2.31.0
Appium-Python-Client~=3.1.1
pytest-bdd~=7.0.1
selenium~=4.17.2
pytest-html==3.2.0